m99.sh: Fix invalid bash command that was causing bad behaviour
The old bash command was not passing the device name to the
m99_latency binary and such argument is mandatory so the binary
cannot run.

Now, the test is designed in the same way as m72 module. The binary
is run in background and after 10 seconds, the process is kill