disable can local loopback test for BL51 board
The BL51 has an issue receiving interrups on the Z029 IP
core so it cannot receive the CAN frames when running
the local loopback test. This is documented in an
Errata document for this Box PC.

Besides, this test for IP core Z029 is beeing performed on a
G229 multifunction board.